+++ src/components/index.ts
@@ -1,13 +1,6 @@
-//컴포넌트 생성시마다 작성된 형식으로 export 한 줄 추가해주세요!å
 export * from 'components/App'
-export * from 'components/Footer'
-export * from 'components/Header'
-export * from 'components/ErrorComponent'
-export * from 'components/NotFound'
-export * from 'components/Modal'
 export * from 'components/admin'
 export * from 'components/mypage'
-export * from 'components/Badge'
-export * from 'components/RecentlyList'
-export * from 'components/Products'
 export * from 'components/product'
+export * from 'components/main'
+export * from 'components/common'
