+++ src/components/main/ImageSlider.tsx
@@ -1,28 +1,28 @@
-import React, { useState } from "react";
-import "../styles/layout/ImageSlider.scss";
+import React, { useState } from 'react'
+import 'styles/layout/ImageSlider.scss'
 
 interface SliderProps {
-  sliderImages: string[];
+  sliderImages: string[]
 }
 
 const ImageSlider: React.FC<SliderProps> = ({ sliderImages = [] }) => {
-  const [currentImage, setCurrentImage] = useState(0);
+  const [currentImage, setCurrentImage] = useState(0)
 
   const goToPreviousImage = () => {
     if (currentImage === 0) {
-      setCurrentImage(sliderImages.length - 1);
+      setCurrentImage(sliderImages.length - 1)
     } else {
-      setCurrentImage(currentImage - 1);
+      setCurrentImage(currentImage - 1)
     }
-  };
+  }
 
   const goToNextImage = () => {
     if (currentImage === sliderImages.length - 1) {
-      setCurrentImage(0);
+      setCurrentImage(0)
     } else {
-      setCurrentImage(currentImage + 1);
+      setCurrentImage(currentImage + 1)
     }
-  };
+  }
 
   return (
     <div className="ImageSlider">
@@ -34,15 +34,19 @@ const ImageSlider: React.FC<SliderProps> = ({ sliderImages = [] }) => {
             className="active"
           />
         )}
-        <button className="ImageSlider_Button_Left" onClick={goToPreviousImage}>
+        <button
+          className="ImageSlider_Button_Left"
+          onClick={goToPreviousImage}>
           <span className="material-symbols-outlined">arrow_back_ios</span>
         </button>
-        <button className="ImageSlider_Button_Right" onClick={goToNextImage}>
+        <button
+          className="ImageSlider_Button_Right"
+          onClick={goToNextImage}>
           <span className="material-symbols-outlined">arrow_forward_ios</span>
         </button>
       </div>
     </div>
-  );
-};
+  )
+}
 
-export { ImageSlider };
+export { ImageSlider }
