+++ src/components/main/index.ts
@@ -1,0 +1,7 @@
+export * from 'components/main/BestSeller'
+export * from 'components/main/ColleyNews'
+export * from 'components/main/ImageSlider'
+export * from 'components/main/NewArrival'
+export * from 'components/main/PromotionSlider'
+export * from 'components/main/RecentlyList'
+export * from 'components/main/Badge'
