+++ src/components/mypage/index.ts
@@ -1,4 +1,3 @@
-//컴포넌트 생성시마다 작성된 형식으로 export 한 줄 추가해주세요!å
 export * from 'components/mypage/MyPageNav'
 export * from 'components/mypage/MyOrderSummary'
 export * from 'components/mypage/MyOrderStatus'
