+++ src/components/product/index.ts
@@ -1,2 +1,3 @@
 export * from 'components/product/ProductDetailInfoTab'
 export * from 'components/product/BestProduct'
+export * from 'components/product/Products'
