+++ src/constants/index.ts
@@ -1,3 +1,4 @@
 export * from 'constants/tag'
 export * from 'constants/errors'
 export * from 'constants/detailTabItems'
+export * from 'constants/imageUrls'
