+++ src/contexts/index.ts
@@ -8,5 +8,4 @@ export * from 'contexts/LoginedUserContext'
 export * from 'contexts/CartContext'
 export * from 'contexts/WishListContext'
 export * from 'contexts/ClickedContext'
-export * from 'contexts/CartLoginedContext'
 export * from 'contexts/UserAddressContext'
