--- src/pages/About.tsx
@@ -1,3 +1,0 @@
-export const About = () => {
-  return <div>About</div>;
-};
