+++ src/pages/index.ts
@@ -1,19 +1,6 @@
-//페이지도 마찬가지로 생성시마다 export추가해주세요
+export * from 'pages/login'
+export * from 'pages/product'
+export * from 'pages/mypage'
+export * from 'pages/payment'
+export * from 'pages/admin'
 export * from 'pages/Home'
-export * from 'pages/About'
-export * from 'pages/SignInPage'
-export * from 'pages/SignUpPage'
-export * from 'pages/AdminDashboard'
-export * from 'pages/AdminProducts'
-export * from 'pages/AdminProductAdd'
-export * from 'pages/ProductList'
-export * from 'pages/ProductDetail'
-export * from 'pages/AdminCustomers'
-export * from 'pages/Payment'
-export * from 'pages/AdminSales'
-export * from 'pages/MyOrders'
-export * from 'pages/Order'
-export * from 'pages/Cart'
-export * from 'pages/Success'
-export * from 'pages/WishList'
-export * from 'pages/ModifyPassword'
