+++ src/pages/mypage/index.ts
@@ -1,0 +1,5 @@
+export * from 'pages/mypage/ModifyPassword'
+export * from 'pages/mypage/MyOrders'
+export * from 'pages/mypage/Order'
+export * from 'pages/mypage/WishList'
+export * from 'pages/mypage/Cart'
