+++ src/pages/payment/index.ts
@@ -1,0 +1,2 @@
+export * from 'pages/payment/Payment'
+export * from 'pages/payment/Success'
