+++ src/pages/product/index.ts
@@ -1,0 +1,2 @@
+export * from 'pages/product/ProductDetail'
+export * from 'pages/product/ProductList'
