+++ src/styles/components/main/recentlyList.module.scss
